% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{annotate}
\alias{annotate}
\title{Annotate}
\usage{
annotate(
  plotlist,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ncol = 1
)
}
\arguments{
\item{plotlist}{list of ggplots}

\item{title}{common title}

\item{xlim}{vector of two numeric values delimiting common x axis}

\item{ylim}{vector of two numeric values delimiting common y axis}

\item{xlab}{common x axis label}

\item{ylab}{common y axis label}

\item{ncol}{number of columns}
}
\value{
annotated figure
}
\description{
Arrange ggplots into grid and annotate the grid.
}
\examples{
library(customplot)
p1 <- scatter_plot(iris, "Sepal.Width", "Sepal.Width", color = "Species")
p2 <- scatter_plot(iris, "Sepal.Width", "Sepal.Length", color = "Species")
p3 <- scatter_plot(iris, "Sepal.Width", "Petal.Width", color = "Species")
p4 <- scatter_plot(iris, "Sepal.Width", "Petal.Length", color = "Species")
annotate(list(p1, p2, p3, p4), title = "iris", xlim = c(0, 5), ylim = 0, 8),
xlab = "Sepal.Width", ylab = "floating", ncol = 2)
}
