% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{between}
\alias{between}
\title{Between}
\usage{
between(value, min, max)
}
\arguments{
\item{value}{controled numeric}

\item{min}{lower limit}

\item{max}{upper limit}
}
\value{
numeric value
}
\description{
Control a numeric value between two numeric limits
}
\examples{
library(customplot)
between(-0.2, 0, 1)
# 0
between(3, 0, 1)
# 1
between(0.75, 0, 1)
# 0.75
}
