% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_adjusted_coords}
\alias{get_adjusted_coords}
\title{Control Legend Coordinates}
\usage{
get_adjusted_coords(coords, hjust, vjust)
}
\arguments{
\item{coords}{vector of two numerics between 0 and 1.}

\item{hjust}{horizontal numeric adjustment}

\item{vjust}{vertical numeric adjustment}
}
\value{
vector of two coordinates
}
\description{
Control coordinates of legend between lower (0) and upper limit (1).
}
\examples{
library(customplot)
coords <- c(1, 1)  # topright
get_adjusted_coords(coords, hjust = -0.1, vjust = -0.1)
# -0.9 -0.9
get_adjusted_coords(coords, hjust = 0.3, vjust = 0.2)
# 1.0 1.0
}
