% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get Color Scheme}
\usage{
get_color_scheme(color_column)
}
\arguments{
\item{color_column}{input vector. If class is factor then color scheme is
6-point brewer. If class is numeric then color scheme is continuous
three-point gradient with mean value in the middle.}
}
\description{
Returns a gg color scheme based on input vector class.
}
\keyword{internal}
