% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_custom_theme}
\alias{get_custom_theme}
\title{Get Custom Theme}
\usage{
get_custom_theme(
  legend,
  angle_text_x,
  angle_text_y,
  legend_hshift,
  legend_vshift,
  legend_box = "vertical",
  legend_direction = "vertical"
)
}
\description{
Helper function.
}
\keyword{internal}
