% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Get Ellipse}
\usage{
get_ellipse(color_column, ellipse)
}
\arguments{
\item{color_column}{numeric vector determining ellipse shape}

\item{ellipse}{boolean enabling ellipse display}
}
\value{
ggplot2::stat_ellipse()
}
\description{
95% t-distribution ellipse corresponding with point color
}
\examples{
library(customplot)
get_ellipse(iris[, "Sepal.Width"], TRUE)
# ggplot2::stat_ellipse(level = 0.95, type = "t")
get_ellipse(iris[, "Sepal.Width"], FALSE)
# NULL
}
