% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_label}
\alias{get_label}
\title{Get Label}
\usage{
get_label(label, overlap)
}
\arguments{
\item{label}{name of the column for point labels; character string. The
column must be numeric.}

\item{overlap}{if FALSE then point label overlapping is avoided. This
action can take more computation time.}
}
\description{
Returns gg label element.
}
\keyword{internal}
