% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_legend_coordinates}
\alias{get_legend_coordinates}
\title{Get Legend Coordinates}
\usage{
get_legend_coordinates(legend, hjust, vjust)
}
\arguments{
\item{legend}{position of a legend; "top", "bottom", "left", "right",
"topleft", "bottomleft", "topright", "bottomright", "center"}

\item{hjust}{legend horizontal shift; numeric value between 0 (very left)
and 1 (very right).}

\item{vjust}{legend vertical shift; numeric value between 0 (very bottom)
and 1 (very top).}
}
\description{
Returns two-element numeric vector of legend coordinates.
}
\keyword{internal}
