% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_lines}
\alias{get_lines}
\title{Get Lines}
\usage{
get_lines(allowed)
}
\arguments{
\item{allowed}{if TRUE then the lines are retained in the output list}
}
\description{
Returns list of gg dashed lines at origin of x- and y-axis.
}
\keyword{internal}
