% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_text_repel}
\alias{get_text_repel}
\title{Get Repel}
\usage{
get_text_repel(label)
}
\arguments{
\item{label}{boolean indicating repel presence}
}
\description{
Positions of labels avoiding an overlap
}
\examples{
library(customplot)
get_text_repel(TRUE)
}
