% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_xlim}
\alias{get_xlim}
\title{Get xlim}
\usage{
get_xlim(xlim)
}
\arguments{
\item{xlim}{vector of two numerics delimiting x axis}
}
\description{
Adjust an x axis proportion.
}
\examples{
library(customplot)
x <- c(0, 22)
get_xlim(x)
}
