% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_ylim}
\alias{get_ylim}
\title{Get ylim}
\usage{
get_ylim(ylim)
}
\arguments{
\item{ylim}{vector of two numerics delimiting y axis}
}
\description{
Adjust an y axis proportion.
}
\examples{
library(customplot)
y <- c(0, 22)
get_xlim(y)
}
