% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot}
\usage{
scatter_plot(
  data_frame,
  x,
  y,
  color = NULL,
  shape = NULL,
  size = NULL,
  label = NULL,
  xlim = NULL,
  ylim = NULL,
  lines = FALSE,
  legend = "topright",
  axis_text_x = TRUE,
  axis_text_y = TRUE,
  ellipse = FALSE,
  point_alpha = 1,
  legend_hshift = 0,
  legend_vshift = 0,
  text_overlap = TRUE,
  legend_box = "vertical",
  legend_direction = "vertical"
)
}
\arguments{
\item{data_frame}{source (data.frame)}

\item{x}{column name (character) determining x coordinates}

\item{y}{column name (character) determining y coordinates}

\item{color}{column name (character) determining point color}

\item{shape}{column name (character) determining point shape}

\item{size}{column name (character) determining point size}

\item{label}{column name (character) determining point labels}

\item{lines}{boolean to highlight axis origin}

\item{legend}{character description of legend position.}

\item{axis_text_x}{boolean to show axis text x}

\item{axis_text_y}{boolean to show axis text y}

\item{ellipse}{boolean to show 95 percent t-distribution ellipse}

\item{point_alpha}{numeric between 0 and 1 determining point transparency}

\item{legend_hshift}{numeric between -1 and 1 determining legend horizontal shift}

\item{legend_vshift}{numeric between -1 and 1 determining legend vertical shift}

\item{text_overlap}{boolean allowing text overlap}

\item{legend_box}{multiple legend orientation; string ("horizontal",
"vertical")}

\item{legend_direction}{legend direction; string ("horizontal", "vertical")}
}
\value{
ggplot
}
\description{
Customized scatter ggplot
}
\examples{
library(customplot)
# legend: "top", "left", "right", "bottom", "center", "topleft", "topright",
# "bottomleft", "bottomright"
scatter_plot(data_frame = iris, x = "Sepal.Width", y = "Sepal.Length",
color = "Petal.Width", size = "Petal.Length", shape = "Species",
label = "Species", lines = TRUE, legend = "right", xaxt = TRUE,
yaxt = FALSE, ellipse = FALSE, alpha = 0.5, hjust = -0.05, vjust = 0.1)
}
