% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{annotate}
\alias{annotate}
\title{Annotate}
\usage{
annotate(
  plotlist,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ncol = 1
)
}
\arguments{
\item{plotlist}{a list of the ggplots}

\item{title}{a title of an annotated figure; character string}

\item{xlim}{a two-element numeric vector delimiting x-axis; numeric vector}

\item{ylim}{a two-element numeric vector delimiting y-axis; numeric vector}

\item{xlab}{x-axis label common for all plots in the grid; character string}

\item{ylab}{y-axis label common for all plots in the grid; character string}

\item{ncol}{a number of the columns in the grid; integer}
}
\value{
the ggplots arranged into an annotated grid
}
\description{
Arranges the ggplots into a grid and annotate it. The number of the grid
columns is user-defined. The number of the rows are recalculated based on
the plotlist length.
}
\examples{
library(customplot)
p1 <- scatter_plot(iris, "Sepal.Width", "Sepal.Width", color = "Species")
p2 <- scatter_plot(iris, "Sepal.Width", "Sepal.Length", color = "Species")
p3 <- scatter_plot(iris, "Sepal.Width", "Petal.Width", color = "Species")
p4 <- scatter_plot(iris, "Sepal.Width", "Petal.Length", color = "Species")
annotate(plotlist = list(p1, p2, p3, p4),
         title = "iris",
         xlim = c(0.5, 5.1),
         ylim = -0.4, 8.3),
         xlab = "Sepal.Width",
         ylab = "floating",
         ncol = 2)
}
