% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_adjusted_coords}
\alias{get_adjusted_coords}
\title{Control Legend Coordinates}
\usage{
get_adjusted_coords(coordinates, legend_hshift, legend_vshift)
}
\arguments{
\item{coordinates}{input legend coordinates; numeric (c(0, 0), c(1, 1))}

\item{legend_hshift}{legend horizontal shift; numeric (-1, 1)}

\item{legend_vshift}{legend vertical shift; numeric (-1, 1)}
}
\value{
vector of two coordinates
}
\description{
support function.
}
\examples{
library(customplot)
coords <- c(1, 1)  # topright
get_adjusted_coords(coords, hjust = -0.1, vjust = -0.1)
# -0.9 -0.9
get_adjusted_coords(coords, hjust = 0.3, vjust = 0.2)
# 1.0 1.0
}
