% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_axis_text}
\alias{get_axis_text}
\title{Get Axis Text}
\usage{
get_axis_text(angle)
}
\arguments{
\item{angle}{angle of tics labels at axis; numeric (0, 180, NULL)}
}
\value{
ggplot2::element_text, ggplot2::element_blank
}
\description{
support function
}
\examples{
library(customplot)
get_axis_text(TRUE)
# ggplot2::element_text()
get_axis_text(FALSE)
# ggplot2::element_blank()
}
