% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_axis_text}
\alias{get_axis_text}
\title{Get Axis Text}
\usage{
get_axis_text(direction)
}
\arguments{
\item{direction}{direction of the text; "horizontal", "vertical", "oblique",
NULL}
}
\description{
Returns the text at ggplot axis
}
\keyword{internal}
