% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_axis_text}
\alias{get_axis_text}
\title{Get Axis Text}
\usage{
get_axis_text(axis_text_visible)
}
\arguments{
\item{axis_text_visible}{boolean value determining returned element}
}
\value{
ggplot2::element_text (or _blank)
}
\description{
Get normal axis text or blank
}
\examples{
library(customplot)
get_axis_text(TRUE)
# ggplot2::element_text()
get_axis_text(FALSE)
# ggplot2::element_blank()
}
