% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Get Color Scheme}
\usage{
get_color_scheme(color_column)
}
\arguments{
\item{color_column}{a single factor or numeric column from a data.frame determining the color}
}
\value{
ggplot2::scale_color_brewer or _gradient2 respectivelly
}
\description{
Get color scale for continuous or categorical scale
}
\examples{
library(customplot)
get_color_scheme(iris[, "Species"])
# ggplot2::scale_color_brewer(palette = "Set2")
get_color_scheme(iris[, "Sepal.Width"])
# ggplot2::scale_color_gradient2(low = "#00AFBB", mid = "#E7B800",
high = "#FC4E07", midpoint = 3.057333)
}
