% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_custom_theme}
\alias{get_custom_theme}
\title{Get Custom Theme}
\usage{
get_custom_theme(
  legend,
  axis_text_x = FALSE,
  axis_text_y,
  legend_hshift = 0,
  legend_vshift,
  legend_box = "vertical",
  legend_direction = "vertical"
)
}
\arguments{
\item{legend}{legend position; string ("top", "left", "right", "bottom",
"center", "topleft", "topright", "bottomleft", "bottomright")}

\item{axis_text_x}{tics labels at x-axis; boolean}

\item{axis_text_y}{tics labels at y-axis; boolean}

\item{legend_hshift}{legend horizontal shift; numeric (-1, 1)}

\item{legend_vshift}{legend vertical shift; numeric (-1, 1)}

\item{legend_box}{multiple legend orientation; string ("horizontal",
"vertical")}

\item{legend_direction}{legend direction; string ("horizontal", "vertical")}
}
\value{
ggplot2::theme()
}
\description{
Customized ggplot theme, text size 8, transparent legend with custom
position, presence/absence of axis text.
}
\examples{
library(customplot)
p1 <- scatter_plot(iris, "Sepal.Width", "Sepal.Length")
p1 + get_custom_theme(legend = "topleft", xaxt = TRUE, yaxt = FALSE,
hjust = 0.1, vjust = -0.1)
}
