% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_custom_theme}
\alias{get_custom_theme}
\title{Get Custom Theme}
\usage{
get_custom_theme(
  legend,
  legend_hshift,
  legend_vshift,
  legend_box,
  legend_direction,
  legend_key_size,
  x_text_direction,
  y_text_direction
)
}
\arguments{
\item{legend}{position of a legend; "top", "bottom", "left", "right",
"topleft", "bottomleft", "topright", "bottomright", "center"}

\item{legend_hshift}{legend horizontal shift; numeric value between 0
(very left) and 1 (very right).}

\item{legend_vshift}{legend vertical shift; numeric value between 0
(very bottom) and 1 (very top).}

\item{legend_box}{orientation of multiple legend (if more than one point
proportion is defined); "horizontal", "vertical"}

\item{legend_direction}{direction of legend keys; "horizontal", "vertical"}

\item{legend_key_size}{legend key size in text lines; numeric value}

\item{x_text_direction}{axis text x direction; "horizontal", "vertical",
"oblique", NULL}

\item{y_text_direction}{axis text y direction; "horizontal", "vertical",
"oblique", NULL}
}
\description{
Returns gg theme
}
\keyword{internal}
