% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_custom_theme}
\alias{get_custom_theme}
\title{Get Custom Theme}
\usage{
get_custom_theme(legend, xaxt, yaxt, hjust, vjust)
}
\arguments{
\item{legend}{character description of legend position. Possible values: "top", "left", "right", "bottom", "center", "topleft", "topright", "bottomleft", "bottomright"}

\item{xaxt}{boolean to show axis text x}

\item{yaxt}{boolean to show axis text y}

\item{hjust}{numeric between -1 and 1 determining legend horizontal shift}

\item{vjust}{numeric between -1 and 1 determining legend vertical shift}
}
\value{
ggplot2::theme()
}
\description{
Customized ggplot theme, text size 8, transparent legend with custom
position, presence/absence of axis text.
}
\examples{
library(customplot)
p1 <- scatter_plot(iris, "Sepal.Width", "Sepal.Length")
p1 + get_custom_theme(legend = "topleft", xaxt = TRUE, yaxt = FALSE,
hjust = 0.1, vjust = -0.1)
}
