% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_label}
\alias{get_label}
\title{Get Label}
\usage{
get_label(label, overlap)
}
\arguments{
\item{label}{column name indicating label presence}

\item{overlap}{boolean allowing text overlap}
}
\description{
Positions of labels avoiding an overlap
}
\examples{
library(customplot)
get_text_repel(TRUE)
}
