% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_legend_coordinates}
\alias{get_legend_coordinates}
\title{Get Legend Coordinates}
\usage{
get_legend_coordinates(legend, hjust, vjust)
}
\arguments{
\item{legend}{character description of legend position. Possible values:
"top", "left", "right", "bottom", "center", "topleft", "topright",
"bottomleft", "bottomright"}

\item{hjust}{numeric between -1 and 1 determining legend horizontal shift}

\item{vjust}{numeric between -1 and 1 determining legend vertical shift}
}
\value{
vector of two numerics
}
\description{
Translate character description of legend position to vector of two
coordinates. Can be adjusted.
}
\examples{
library(customplot)
get_legend_coordinates("left", hjust = 0.1, vjust = 0.1)
# ggplot2::stat_ellipse(level = 0.95, type = "t")
get_ellipse(iris[, "Sepal.Width"], FALSE)
# NULL
}
