% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_lines}
\alias{get_lines}
\title{Get Lines}
\usage{
get_lines(lines)
}
\arguments{
\item{lines}{boolean showing axis origin}
}
\value{
list of ggplot2::geom_hline() and _vline()
}
\description{
Highlight axis x and y at their origin
}
\examples{
library(customplot)
get_lines(TRUE)
# [[1]]
# mapping: xintercept = ~xintercept
# geom_vline: na.rm = FALSE
# stat_identity: na.rm = FALSE
# position_identity
# [[2]]
# mapping: yintercept = ~yintercept
# geom_hline: na.rm = FALSE
# stat_identity: na.rm = FALSE
# position_identity
}
