% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{get_mapping}
\alias{get_mapping}
\title{Get Mapping}
\usage{
get_mapping(x, y, color, shape, size, label)
}
\arguments{
\item{x}{column name (character) determining x coordinates}

\item{y}{column name (character) determining y coordinates}

\item{color}{column name (character) determining point color}

\item{shape}{column name (character) determining point shape}

\item{size}{column name (character) determining point size}

\item{label}{column name (character) determining point labels}
}
\value{
ggplot2::aes()
}
\description{
Calculate the proportions of a ggplot
}
\examples{
library(customplot)
get_mapping(x = "Sepal.Width", y = "Sepal.Length", color = "Petal.Width",
size = "Petal.Length", shape = "Species", label = "Species")
}
