% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot}
\usage{
scatter_plot(
  data_frame,
  x,
  y,
  color = NULL,
  shape = NULL,
  size = NULL,
  label = NULL,
  xlim = NULL,
  ylim = NULL,
  lines = FALSE,
  legend = "topright",
  xaxt = TRUE,
  yaxt = TRUE,
  ellipse = FALSE,
  alpha = 1,
  hjust = 0,
  vjust = 0,
  overlap = TRUE
)
}
\arguments{
\item{data_frame}{source (data.frame)}

\item{x}{column name (character) determining x coordinates}

\item{y}{column name (character) determining y coordinates}

\item{color}{column name (character) determining point color}

\item{shape}{column name (character) determining point shape}

\item{size}{column name (character) determining point size}

\item{label}{column name (character) determining point labels}

\item{lines}{boolean to highlight axis origin}

\item{legend}{character description of legend position.}

\item{xaxt}{boolean to show axis text x}

\item{yaxt}{boolean to show axis text y}

\item{ellipse}{boolean to show 95 percent t-distribution ellipse}

\item{alpha}{numeric between 0 and 1 determining point transparency}

\item{hjust}{numeric between -1 and 1 determining legend horizontal shift}

\item{vjust}{numeric between -1 and 1 determining legend vertical shift}

\item{overlap}{boolean allowing text overlap}
}
\value{
ggplot
}
\description{
Customized scatter ggplot
}
\examples{
library(customplot)
# legend: "top", "left", "right", "bottom", "center", "topleft", "topright",
# "bottomleft", "bottomright"
scatter_plot(data_frame = iris, x = "Sepal.Width", y = "Sepal.Length",
color = "Petal.Width", size = "Petal.Length", shape = "Species",
label = "Species", lines = TRUE, legend = "right", xaxt = TRUE,
yaxt = FALSE, ellipse = FALSE, alpha = 0.5, hjust = -0.05, vjust = 0.1)
}
