% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot}
\usage{
scatter_plot(
  data_frame,
  x,
  y,
  color = NULL,
  shape = NULL,
  size = NULL,
  label = NULL,
  xlim = NULL,
  ylim = NULL,
  lines = FALSE,
  legend = "topright",
  x_text_direction = "horizontal",
  y_text_direction = "horizontal",
  ellipse = FALSE,
  point_alpha = 1,
  legend_hshift = 0,
  legend_vshift = 0,
  text_overlap = TRUE,
  legend_box = "vertical",
  legend_direction = "vertical",
  legend_key_size = 1
)
}
\arguments{
\item{data_frame}{source data frame}

\item{x}{name of the column containing x-coordinates; character string.
The column must be factor or numeric.}

\item{y}{name of the column containing y-coordinates; character string.
The column must be factor or numeric.}

\item{color}{name of the column for point color values; character string.
The column must be factor or numeric.}

\item{shape}{name of the column for point shape values; character string.
The column must be factor.}

\item{size}{name of the column for point size values; character string.
The column must be numeric.}

\item{label}{name of the column for point labels; character string. The
column must be character.}

\item{lines}{if TRUE then origin is highlighted by dashed x- and y-lines}

\item{legend}{position of a legend; "top", "bottom", "left", "right",
"topleft", "bottomleft", "topright", "bottomright", "center"}

\item{x_text_direction}{axis text x direction; "horizontal", "vertical",
"oblique", NULL}

\item{y_text_direction}{axis text y direction; "horizontal", "vertical",
"oblique", NULL}

\item{ellipse}{if TRUE and color column is factor then 95-percent
t-distribution ellipses appear around the color groups.}

\item{point_alpha}{point transparency; numeric value between 0
(fully transparent) and 1 (fully solid).}

\item{legend_hshift}{legend horizontal shift; numeric value between 0
(very left) and 1 (very right).}

\item{legend_vshift}{legend vertical shift; numeric value between 0
(very bottom) and 1 (very top).}

\item{text_overlap}{if FALSE then point label overlapping is avoided. This
action can take more computation time.}

\item{legend_box}{orientation of multiple legend (if more than one point
proportion is defined); "horizontal", "vertical"}

\item{legend_direction}{direction of legend keys; "horizontal", "vertical"}

\item{legend_key_size}{legend key size in text lines; numeric value}
}
\value{
customized ggplot
}
\description{
Returns a figure of a customized scatter ggplot.
}
\examples{
library(customplot)
scatter_plot(data_frame = iris, x = "Sepal.Width", y = "Sepal.Length",
color = "Petal.Width", size = "Petal.Length", shape = "Species",
label = "Species", lines = TRUE, legend = "right", point_alpha = 0.5,
legend_hshift = -0.05, legend_vshift = 0.1)
}
